systeme de deconnexion, changement co / deco

+++ suppression-msg.php
@@ -1,14 +1,16 @@
 <?php
 	include('includes/connexion.inc.php');
 
-		
-	/*Vérfication de l'ID pour préparer l'update*/
-	if(isset($_GET['id']) && !empty($_GET['id'])){
-	$query = 'DELETE FROM messages WHERE id = (:id)';
-	$prep = $pdo->prepare($query);
-	$prep->bindValue(':id',$_GET['id']);
-	$prep->execute();
+	if($connecte == true){
+		/*Vérfication de l'ID pour préparer l'update*/
+		if(isset($_GET['id']) && !empty($_GET['id'])){
+			$query = 'DELETE FROM messages WHERE id = (:id)';
+			$prep = $pdo->prepare($query);
+			$prep->bindValue(':id',$_GET['id']);
+			$prep->execute();
+		}
 	}
+
